+++ inscripta/biocantor/io/genbank/parser.py
@@ -74,7 +74,7 @@ class Feature(ABC):
     def __init__(self, feature: SeqFeature, record: SeqRecord):
         if feature.type not in self.types:
             raise GenBankParserError(f"Invalid feature type {feature.type}")
-        if not feature.location:
+        if feature.location is None:
             raise GenBankLocationException(f"Feature {feature} did not have parseable coordinates.")
         if not feature.strand:
             raise GenBankParserError(f"Feature {feature} is unstranded or has multiple strands.")
@@ -117,7 +117,7 @@ class FeatureIntervalGenBankCollection:
             record: The ``SeqRecord`` these features were found on.
         """
         for feature in features:
-            if not feature.location:
+            if feature.location is None:
                 raise GenBankLocationException(f"Feature {feature} did not have parseable coordinates.")
 
         self.types = {feature.type for feature in features}
@@ -215,7 +215,10 @@ class GeneFeature(Feature):
         return self.feature.__repr__()
 
     @staticmethod
-    def from_transcript_feature(feature: SeqFeature, seqrecord: SeqRecord) -> "GeneFeature":
+    def from_transcript_or_cds_feature(feature: SeqFeature, seqrecord: SeqRecord) -> "GeneFeature":
+        """Some GenBank files lack a gene-level feature, but have transcript-level features or CDS-level features only.
+
+        Construct a GeneFeature from such records."""
         old_type = feature.type
         feature.type = "gene"
         gene = GeneFeature(feature, seqrecord)
@@ -299,7 +302,7 @@ class GeneFeature(Feature):
                 cds_starts=cds_starts,
                 cds_ends=cds_ends,
                 cds_frames=cds_frames,
-                qualifiers=tx.feature.qualifiers,
+                qualifiers=tx.merge_cds_qualifiers_to_transcript(),
                 is_primary_tx=False,
                 transcript_id=tx.get_qualifier_from_tx_or_cds_features("transcript_id"),
                 protein_id=tx.get_qualifier_from_tx_or_cds_features("protein_id"),
@@ -396,6 +399,21 @@ class TranscriptFeature(Feature):
         )
         return exon_interval.intersection(cds_i)
 
+    def merge_cds_qualifiers_to_transcript(self) -> Dict[str, List[str]]:
+        """
+        If there were distinct transcript-level features, the qualifiers on the CDS feature will be lost
+        when converting to the BioCantor data model unless those qualifiers are rolled into the qualifiers on
+        the transcript feature.
+        """
+        qualifiers = {key: set(vals) for key, vals in self.feature.qualifiers.items()}
+        for cds_feature in self.cds_features:
+            for key, vals in cds_feature.feature.qualifiers.items():
+                if key not in qualifiers:
+                    qualifiers[key] = set(vals)
+                else:
+                    qualifiers[key].update(vals)
+        return {key: list(vals) for key, vals in qualifiers.items()}
+
 
 class IntervalFeature(Feature):
     """A set of intervals"""
@@ -441,6 +459,16 @@ def group_gene_records_from_sorted_genbank(
 ) -> Iterable[ParsedAnnotationRecord]:
     """Model 1: position sorted GenBank.
 
+    This function looks for canonical gene records:
+        gene -> Optional(mRNA) -> CDS records
+    It also looks for canonical non-coding records:
+        gene -> {misc_RNA,tRNA,rRNA,etc)
+
+    It also will infer non-canonical record types, including non-coding transcripts and coding genes
+    from isolated CDS/non-coding features (those without a gene feature before them in the sort order).
+
+    Any features that do not fit the above bins are interpreted as generic features.
+
     Args:
         record_iter: Iterable of SeqRecord objects.
         parse_func: Optional parse function implementation.
@@ -465,9 +493,19 @@ def group_gene_records_from_sorted_genbank(
             elif gene is None:
                 if feature.type in GeneFeature.types:
                     gene = GeneFeature(feature, seqrecord)
+                # base case for starting with a isolated ncRNA or CDS feature; immediately add them
+                # and reset the gene to None
+                elif feature.type in TranscriptFeature.types or feature.type in IntervalFeature.types:
+                    gene = GeneFeature.from_transcript_or_cds_feature(feature, seqrecord)
+                    gene.finalize()
+                    gene = parse_func(gene)
+                    genes.append(gene)
+                    gene = None
+                # this must be a generic feature
                 else:
-                    continue
-            elif feature.type in GeneFeature.types:  # next gene
+                    feature_features.append(feature)
+            # next gene; re-set the gene object and report out the collection
+            elif feature.type in GeneFeature.types:
                 if gene.has_children:
                     gene.finalize()
                     gene = parse_func(gene)
@@ -480,11 +518,11 @@ def group_gene_records_from_sorted_genbank(
                     gene.finalize()
                     gene = parse_func(gene)
                     genes.append(gene)
-                    gene = GeneFeature.from_transcript_feature(feature, seqrecord)
+                    gene = GeneFeature.from_transcript_or_cds_feature(feature, seqrecord)
                 else:
                     gene.add_child(feature)
             elif feature.type in IntervalFeature.types:
-                if len(gene.children) == 0:
+                if not gene.has_children:
                     gene.add_child(feature)
                 else:
                     gene.children[-1].add_child(feature)
@@ -563,7 +601,8 @@ def group_gene_records_by_locus_tag(
         for locus_tag, gene_features in itertools.groupby(
             sorted_gene_filtered_features, key=lambda f: f.qualifiers["locus_tag"][0]
         ):
-            gene_features = list(gene_features)
+            # sort the features for this locus tag to bubble the "gene" feature to the top, if it exists
+            gene_features = sorted(gene_features, key=lambda f: f.type != "gene")
             gene = gene_features[0]
 
             if gene.type not in GeneFeature.types: